// src/api/generated/zod.ts
export const putAssumptionsApiV1AssumptionsPutBodyHurdleFixedRateRegExpTwo = new RegExp('^(?!^[-+.]\*$)[+-]?0\*\\d\*\\.?\\d\*$');


export const putAssumptionsApiV1AssumptionsPutBody = zod.object({
  "display_currency": zod.enum(['irr', 'usd', 'both']).optional(),
  "hurdle_mode": zod.enum(['fixed', 'inflation', 'usd_growth']).optional(),
  "hurdle_fixed_rate": zod.union([zod.number(),zod.string().regex(putAssumptionsApiV1AssumptionsPutBodyHurdleFixedRateRegExpTwo),zod.null()]).optional(),
  "growth_assumptions": zod.union([zod.record(zod.string(), zod.unknown()),zod.null()]).optional()
})