// src/api/generated/schema.ts
export interface components {
    schemas: {
// ...
        AssumptionsUpdate: {
            /** @default both */
            display_currency: components["schemas"]["DisplayCurrency"];
            /** @default inflation */
            hurdle_mode: components["schemas"]["HurdleMode"];
            /** Hurdle Fixed Rate */
            hurdle_fixed_rate?: number | string | null;
            /** Growth Assumptions */
            growth_assumptions?: {
                [key: string]: unknown;
            } | null;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}